[DEV] organize version directories
