+++ dbrx/b2_opt_ser/serialization_utils.py
@@ -1,6 +1,6 @@
 import io
 
-import numpy as np
+# import numpy as np
 import mlx.core as mx
