--- dbrx/base.py
@@ -1,15 +1,0 @@
-import inspect
-from dataclasses import dataclass
-
-
-@dataclass
-class BaseModelArgs:
-    @classmethod
-    def from_dict(cls, params):
-        return cls(
-            **{
-                k: v
-                for k, v in params.items()
-                if k in inspect.signature(cls).parameters
-            }
-        )
