+++ dbrx/batch1/launch_moe_shards_batch1.py
@@ -1,8 +1,15 @@
+# Examples:
+#  launch all shards:
+#    python launch_moe_shards_batch1.py --model-path ~/dbrx-base/distributable/batch1/
+#
+#  terminate all shards:
+#    python launch_moe_shards_batch1.py --model-path ~/dbrx-base/distributable/batch1/ --terminate
 import subprocess
-import argparse
 import time
 from types import SimpleNamespace
-import sys
+import argparse
+from pathlib import Path
+import json
 
 """terminal color"""
 TC = SimpleNamespace(
@@ -18,7 +25,7 @@ TC = SimpleNamespace(
 
 class Cmd:
     def __new__(
-        self, cmd: str, cwd="./", timeout_duration=None, suppress=False
+        self, cmd: str, cwd="./", timeout_duration=None, suppress=True
     ) -> tuple[int, str, str]:
         self.cmd = cmd
         self.cwd = cwd
@@ -90,45 +97,49 @@ class Cmd:
         return self.returncode, str(out, encoding="utf8"), str(err, encoding="utf8")
 
 
-def list_of_strings(arg):
-    return arg.split(",")
-
-
 def main():
     parser = argparse.ArgumentParser()
-    # parser.add_argument("--ports", required=True, type=str)
-    parser.add_argument("--ports", type=list_of_strings)
-    parser.add_argument("--terminate", action="store_true")
-
-    args = parser.parse_args()
-    ports = args.ports
-
-    print("Sorry, we need to kill the tmux server")
-    Cmd("""tmux kill-server""")
-    if args.terminate:
-        return
-
-    rc, out, err = Cmd(
-        """tmux -f /dev/null new-session -s dbrx_poc -n experts -d zsh \;"""
+    parser.add_argument(
+        "--model-path", required=True, type=str, help="Path to local model directory"
     )
-    if rc != 0:
-        print(err, file=sys.stderr)
-        sys.exit(1)
-    Cmd("""tmux set-option -g mouse on""")
-    Cmd("""tmux split-window -hf zsh \;""")
-    Cmd("""tmux split-window -hf zsh \;""")
-    Cmd("""tmux split-window -hf zsh \;""")
-    Cmd("""tmux select-layout even-horizontal \;""")
-    for e, port in enumerate(ports):
-        Cmd(f"""tmux send-keys -t {e} 'clear' Enter \;""")
-        Cmd(f"""tmux send-keys -t {e} 'conda activate dbrx_poc' Enter \;""")
-        Cmd(f"""tmux send-keys -t {e} 'cd ~/research_muchi/dbrx/' Enter \;""")
+    parser.add_argument("--terminate", action="store_true")
+    args = parser.parse_args()
+    try:
+        with open(Path(args.model_path) / "driver_config.json", "r") as f:
+            config = json.load(f)
+    except FileNotFoundError:
+        raise
+
+    moe_shard_map = config["ffn_config"]["moe_shard_map"]
+    pure_urls = [url.split(":")[0] for url in moe_shard_map.keys()]
+    url_port_map = {url: [] for url in pure_urls}
+    for url, experts in moe_shard_map.items():
+        pure_url, port = url.split(":")
+        url_port_map[pure_url].append(port)
+
+    for url, ports in url_port_map.items():
+        print(f"Shard: {url} {ports} ", end="")
         Cmd(
-            f"""tmux send-keys -t {e} 'python moe_shard_ser.py --port {port}"""
-            + f""" --model-path ~/dbrx-base/distributable/batch2"""
-            + f""" --config-filename moe_shard_config_{e}.json' Enter \;""",
+            f"""scp -i ~/.ssh/id_llamacpp ./run_shard_batch1.py xiangruike@{url}:/users/xiangruike"""
         )
-    # Cmd("""tmux -f /dev/null attach -t dbrx_poc""")
+        if args.terminate:
+            rc, out, err = Cmd(
+                f"""ssh -i ~/.ssh/id_llamacpp xiangruike@{url} 'export PATH="$PATH:/opt/homebrew/bin/" """
+                + f"""&& python3 /Users/xiangruike/run_shard_batch1.py --ports "{','.join(ports)}" --terminate'"""
+            )
+            if rc != 0:
+                print(err.strip())
+            else:
+                print("[terminated successfully]")
+        else:
+            rc, out, err = Cmd(
+                f"""ssh -i ~/.ssh/id_llamacpp xiangruike@{url} 'export PATH="$PATH:/opt/homebrew/bin/" """
+                + f"""&& python3 /Users/xiangruike/run_shard_batch1.py --ports "{','.join(ports)}"'"""
+            )
+            if rc != 0:
+                print(err.strip())
+            else:
+                print("[launched successfully]")
 
 
 if __name__ == "__main__":
