+++ dbrx/batch2/moe_shard_batch2.proto
